+++ R/data-structures.R
@@ -1,14 +1,15 @@
+
+py_list <- function(x) {
+  (.globals$py_list %||%
+     (.globals$py_list <- reticulate::import_builtins()$list))(x)
+}
+
 #' @importFrom reticulate py_to_r
 #' @export
-py_to_r.tensorflow.python.training.tracking.data_structures.List <- function(x) {
-  py_builtin()$list(x)
-}
+py_to_r.tensorflow.python.training.tracking.data_structures.List <- py_list
 
+#' @export
+py_to_r.tensorflow.python.trackable.data_structures.ListWrapper <- py_list
 
-py_builtin <- function() {
-  builtin <- .globals$py_builtin
-  if (is.null(builtin)) {
-    builtin <- .globals$py_builtin <- reticulate::import_builtins()
-  }
-  builtin
-}
+#' @export
+py_to_r.tensorflow.python.trackable.data_structures.List <- py_list